requirements.txt
flask
flask_cors
flask_sqlalchemy
sentence-transformers
torch
pdfplumber
python-docx
fpdf
werkzeug
gunicorn
